+++ modules/architecture-representation/src/main/java/br/ufpr/dinf/gres/architecture/toSMarty/relationship/SaveRequiresSMarty.java
@@ -1,0 +1,59 @@
+package br.ufpr.dinf.gres.architecture.toSMarty.relationship;
+
+import br.ufpr.dinf.gres.architecture.representation.Architecture;
+import br.ufpr.dinf.gres.architecture.representation.Element;
+import br.ufpr.dinf.gres.architecture.representation.relationship.Relationship;
+import br.ufpr.dinf.gres.architecture.representation.relationship.RequiresRelationship;
+import br.ufpr.dinf.gres.architecture.toSMarty.util.SaveStringToFile;
+
+import java.io.PrintWriter;
+
+public class SaveRequiresSMarty {
+
+    public SaveRequiresSMarty(Architecture architecture, PrintWriter printWriter, String logPath) {
+        String halfTab = "  ";
+        String tab = "    ";
+        int id_rel = 1;
+        for (Relationship r : architecture.getRelationshipHolder().getAllRelationships()) {
+            if (r instanceof RequiresRelationship) {
+                // <dependency source="CLASS#12" target="INTERFACE#10"/>
+                RequiresRelationship dr = (RequiresRelationship) r;
+
+                Element e1 = architecture.findElementByNameInPackageAndSubPackage(dr.getClient().getName());
+                if (e1 == null) {
+                    System.out.println("Discart Req 1:" + dr.getClient().getId());
+                    SaveStringToFile.getInstance().appendStrToFile(logPath, "\n\nDiscart Requires " + dr.getId() + ":");
+                    SaveStringToFile.getInstance().appendStrToFile(logPath, "\nSupplier: " + dr.getSupplier().getId() + " - " + dr.getSupplier().getName());
+                    SaveStringToFile.getInstance().appendStrToFile(logPath, "\nClient: " + dr.getClient().getId() + " - " + dr.getClient().getName() + " not found");
+                    continue;
+                }
+                Element e2 = architecture.findElementByNameInPackageAndSubPackage(dr.getSupplier().getName());
+                if (e2 == null) {
+                    System.out.println("Discart Req 2:" + dr.getSupplier().getId());
+                    SaveStringToFile.getInstance().appendStrToFile(logPath, "\n\nDiscart Requires " + dr.getId() + ":");
+                    SaveStringToFile.getInstance().appendStrToFile(logPath, "\nSupplier: " + dr.getSupplier().getId() + " - " + dr.getSupplier().getName() + " not found");
+                    SaveStringToFile.getInstance().appendStrToFile(logPath, "\nClient: " + dr.getClient().getId() + " - " + dr.getClient().getName());
+                    continue;
+                }
+                if (dr.getId().length() == 0) {
+                    boolean existID = true;
+                    while (existID) {
+                        existID = false;
+                        for (Relationship r2 : architecture.getRelationshipHolder().getAllRelationships()) {
+                            if (r2.getId().equals("REQUIRES#" + id_rel)) {
+                                id_rel++;
+                                existID = true;
+                                break;
+                            }
+                        }
+                    }
+                    dr.setId("REQUIRES#" + id_rel);
+                    id_rel++;
+                }
+                printWriter.write("\n" + tab + "<requires id=\"" + dr.getId() + "\" source=\"" + e1.getId() + "\" target=\"" + e2.getId() + "\">");
+                printWriter.write("\n" + tab + "</requires>");
+            }
+        }
+    }
+
+}
