+++ modules/opla-core/src/main/java/br/ufpr/dinf/gres/core/jmetal4/metaheuristics/nsgaII/NSGAII.java
@@ -22,6 +22,8 @@
 package br.ufpr.dinf.gres.core.jmetal4.metaheuristics.nsgaII;
 
 import br.ufpr.dinf.gres.architecture.io.OPLALogs;
+import br.ufpr.dinf.gres.architecture.io.ReaderConfig;
+import br.ufpr.dinf.gres.architecture.toSMarty.util.SaveStringToFile;
 import br.ufpr.dinf.gres.domain.OPLAThreadScope;
 import br.ufpr.dinf.gres.architecture.io.OptimizationInfo;
 import br.ufpr.dinf.gres.architecture.io.OptimizationInfoStatus;
@@ -37,6 +39,9 @@ import br.ufpr.dinf.gres.core.learning.ClassifierAlgorithm;
 import br.ufpr.dinf.gres.core.learning.SubjectiveAnalyzeAlgorithm;
 import org.apache.log4j.Logger;
 
+import java.io.File;
+import java.io.FileWriter;
+import java.io.PrintWriter;
 import java.util.HashSet;
 import java.util.stream.Collectors;
 
@@ -112,6 +117,15 @@ public class NSGAII extends Algorithm {
         selectionOperator = operators_.get("selection");
 
         try {
+
+            Solution solution_base = new Solution(problem_);
+            //Architecture architecture = (Architecture)solution_base.getDecisionVariables()[0];
+            //architecture.save2(architecture,"ServiceAndSupportSystemAtual");
+            //architecture.openTempArch();
+            problem_.evaluateConstraints(solution_base);
+            problem_.evaluate(solution_base);
+            saveBaseHypervolume(solution_base);
+
             LOGGER.info("Criando População");
             // Create the initial solutionSet
             Solution newSolution;
@@ -328,5 +342,24 @@ public class NSGAII extends Algorithm {
 
         problem_.evaluateConstraints(newSolution);
         return newSolution;
+    }
+
+    private void saveBaseHypervolume(Solution solution){
+        SaveStringToFile.getInstance().createLogDir();
+        String path = ReaderConfig.getDirExportTarget()+"/Logs/hypervolume_base.txt";
+        try {
+            FileWriter fileWriter = new FileWriter(path);
+            PrintWriter printWriter = new PrintWriter(fileWriter);
+
+            for(Double fit : solution.getObjectives()) {
+                printWriter.write(fit.toString());
+                printWriter.write(" ");
+            }
+            printWriter.close();
+            fileWriter.close();
+        }catch (Exception ex){
+            System.out.println(ex);
+            ex.printStackTrace();
+        }
     }
 } // NSGA-II
