+++ modules/opla-core/src/main/java/br/ufpr/dinf/gres/core/jmetal4/operators/mutation/PLAFeatureMutation.java
@@ -1169,7 +1169,7 @@ public class PLAFeatureMutation extends Mutation {
                 }
             }
         }
-        variabilities.clear();
+        //variabilities.clear();
         return isVariantConcern;
     }
